set up criwin:1.2

+++ src/commands/download/download_image.py
@@ -48,7 +48,8 @@ async def setup_download_image(interaction: Interaction, active_downloads: set[i
         await send_message(interaction, str(error))
         
     finally:
-        for file in file_path:
-            if file is not None:
-                file.unlink(missing_ok=True)
+        if file_path:
+            for file in file_path:
+                if file is not None:
+                    file.unlink(missing_ok=True)
         active_downloads.discard(user_id)
